+++ src/dipdup/datasources/bcd/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+# TODO: Use in config
+class BcdNetwork(Enum):
+    mainnet = 'mainnet'
+    hangzhou2net = 'hangzhou2net'