+++ tests/test_dipdup/types/listofmaps/storage.py
@@ -1,9 +1,8 @@
 from typing import Dict
 from typing import List
-from typing import Union
 
 from pydantic import BaseModel
 
 
 class ListOfMapsStorage(BaseModel):
-    __root__: List[Union[int, Dict[str, str]]]
+    __root__: List[Dict[str, str]]